Moved translations to subfolder to clean up file structure
