+++ languages/descriptions_hu.py
