+++ languages/descriptions_ko.py
